==> ipynb/1a-get-tweets-got3.ipynb <==
## Notebook 1 - Get Tweets using GetOldTweets
The purpose of this notebook is to gather the tweets. Scott Tarlow, a contact shared with me by my advisor Joshua Cook, directed me to this GetOldTweets-python repo (https://github.com/Jefferson-Henrique/GetOldTweets-python) to circumvent the limits of the Twitter API. Here are some of the tweeks I implemented in GetOldTweets to tailor it to my needs:  
    
**Problem:**  
The functionality was written for Python2 and I'm using Python3.  

**Fix:**  
Replaced `urllib2` references with `urllib` in `getJsonReponse` function within `TweetManager.py`.   
    
**Problem:**  
I also discovered that the more tweets I requested, the time to retrieve them grew exponentially. 

**Fix:**  
Added functionality to pull tweets one month at a time and it does so at a rate of about 1500 tweets per minute (drastic improvement over the 100 minutes it took to pull 500 tweets previously).
  
**Problem:**  
Tweets sent as replies to other tweets were being included. I only want original tweets.  
  
**Fix:**  
I added a check in `getTweets` function within `TweetManager.py` and it skips replies.  
  
**Problem:**  
The original code references a section of the Twitter API repsonse that no longer exists. It led to empty username and location data in the data frame.  
  
**Fix:**  
I relocated where username is in the response, and updated the code. However, I could not find any location information.

cd /home/jovyan/capstone/src/

import pandas as pd

import got3 as got

def get_tweets(username):
    from collections import defaultdict
    d = defaultdict(list)
    
    attributes = ['date', 'favorites', 'retweets', 'hashtags', 'id', 'mentions',
                  'text', 'urls', 'username', 'permalink', 'author_id']

    # created to efficiently pull tweets one month at a time
    years = ['2008', '2009', '2010', '2011', '2012', '2013', '2014', '2015', '2016', '2017']
    months = [('01', '31'), ('02', '28'), ('03', '31'), ('04', '30'), ('05', '31'), ('06', '30'),
              ('07', '31'), ('08', '31'), ('09', '30'), ('10', '31'), ('11', '30'), ('12', '31')]
    total = 0
    for year in years:
        for month in months:
            
            # account for non-inclusive end of month by starting where previous month left off
            if year == '2008' and month[0] == '01':
                since = year + '-' + month[0] + '-01'
            else:
                since = until
            # set the end date to beginning of next month
            until = year + '-' + month[0] + '-' + month[1]
            
            tweetCriteria = got.manager.TweetCriteria().setUsername(username).setSince(since).setUntil(until)
            tweet_list = got.manager.TweetManager.getTweets(tweetCriteria)
            print('{} tweets from {} to {}.'.format(len(tweet_list), since, until), end='\r')
            
            total += len(tweet_list)
            for tweet in tweet_list:
                for att in attributes:
                    d[att].append(eval("tweet." + att))
    
    print('{} total tweets for {}.'.format(total, username))
    return pd.DataFrame(d)

def get_and_export(filename, handle):
    data = get_tweets(str(handle))
    data.to_pickle('../data/1-' + str(filename) + '_tweets.p')

# list of tuples containing filename for pickle and twitter handle
all_spcas = [('sfspca', 'sfspca'), ('pspca', 'PSPCA'), ('houston', 'HoustonSPCA'), ('texas', 'spcaoftexas'),
            ('tulsa', 'Tulsa_SPCA'), ('richmond', 'RichmondSPCA'), ('ontario', 'OntarioSPCA'), 
            ('alberta', 'FMSPCA'), ('bc', 'BC_SPCA')]

count = 0
for filename, handle in all_spcas:
    count += 1
    print('Pulling {} of {}: {}'.format(count, len(all_spcas), handle))
    get_and_export(filename, handle)
