==> ipynb/1b-get-tweets-streaming.ipynb <==
## Notebook 1b - Get Tweets using Twitter Streaming API
The purpose of this notebook is to gather random tweets from across the US and Canada so that I can use them to fit the text encoder.

# !pip install twitter

# !pip install pymongo

CONSUMER_KEY = 
CONSUMER_SECRET = 
ACCESS_TOKEN = 
ACCESS_SECRET = 

from twitter import TwitterStream
from twitter import OAuth

oauth = OAuth(ACCESS_TOKEN,
              ACCESS_SECRET,
              CONSUMER_KEY,
              CONSUMER_SECRET)

# using this box for the eastern half
east_bbox = "-95.05,30.49,-65.0,52.4"

# connect to the API and create iterator with desired location
twitter_stream = TwitterStream(auth=oauth)
twitterator = (twitter_stream
               .statuses
               .filter(locations=east_bbox))

# use this function if writing tweets to MongoDB
from pymongo import MongoClient
def tweets_to_mongo():
    client = MongoClient('54.191.107.201', 27016)
    db_ref = client.twitter
    coll_ref = db_ref.tweets

    i = 1
    while i < 1001:
        coll_ref.insert_one(next(twitterator))
        if i % 100 == 0:
            print('{} tweets written to database.'.format(coll_ref.count()))
        i += 1

# use this function if pickling tweets
import pandas as pd
def tweets_to_file(filename):
    east_tweets = []
    i = 1
    while i < 300001:
        east_tweets.append(next(twitterator))

        # counter showing tweets and the file they'll be written to
        print('tweets: {}, file: {}'.format(i,filename), end='\r')

        # write to the pickle file every 1000 tweets
        if i % 1000 == 0:
            pd.DataFrame(east_tweets).to_pickle(filename)

        # change filename every 10,000 tweets and restart the list
        if i % 10000 == 0:
            
            # account for filenames ending in 01 through 08
            if int(filename[-4:-2]) < 9:
                filename = 'data/east-0{}.p'.format(int(filename[-4:-2])+1)
            else:
                filename = 'data/east-{}.p'.format(int(filename[-4:-2])+1)
            
            east_tweets = []
        
        i += 1

tweets_to_file('data/east-01.p')
